agent/requirements.txt
flask
requests
google-cloud-bigquery
google-cloud-firestore
